conversiondireccion

+++ api2/src/models/safePlace.js
@@ -38,11 +38,11 @@ module.exports = function (sequelize) {
             allowNull:false,
         },
         lat: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.INTEGER,
             allowNull: true,
         },
         lng: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.INTEGER,
             allowNull: true,
         },
         email:{
